Connection to MySQL server now works, will deploy jar at later time

+++ Backend/project/project2/src/main/java/data/User/AppUserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 public interface AppUserRepository extends JpaRepository<AppUser, Long> {
-        AppUser findById(int UserID);
+        AppUser findById(long UserID);
 
         @Transactional
-        void deleteById(int UserID);
+        void deleteById(long UserID);
     }
